src/requirement.txt
pyfiglet
psycopg2-binary
sqlalchemy
jsonpickle == 3.0.3

aioredis == 2.0.1
email_validator == 2.1.1
attrs
requests
mongoengine == 0.28.2
flask_socketio
socketio